Fix HPMultiplier type
최소, 최대 배율 모두 자연수가 아닐 수 있습니다.

+++ internal/types/schaledb.go
@@ -107,10 +107,10 @@ type SkillEffect struct {
 
 // HP-based damage modifier
 type HpRateDamageModifier struct {
-	MaxHpRate     int `json:"MaxHpRate"`
-	MinHpRate     int `json:"MinHpRate"`
-	MultiplierMax int `json:"MultiplierMax"`
-	MultiplierMin int `json:"MultiplierMin"`
+	MaxHpRate     int     `json:"MaxHpRate"`
+	MinHpRate     int     `json:"MinHpRate"`
+	MultiplierMax float64 `json:"MultiplierMax"`
+	MultiplierMin float64 `json:"MultiplierMin"`
 }
 
 // Student weapon
